refactor(building): split extraction.rs into modular directory
Break apart the 740 LOC extraction.rs file into a focused directory structure
where each element has its own file with colocated tests, improving code
organization and developer experience.

Structure:
- extraction/annotation.rs (316 lines) - Annotation and parameter extraction
- extraction/verbatim.rs (200 lines) - Verbatim block with indentation handling
- extraction/paragraph.rs (89 lines) - Paragraph extraction with tests
- extraction/session.rs (42 lines) - Session extraction
- extraction/definition.rs (42 lines) - Definition extraction
- extraction/list_item.rs (42 lines) - List item extraction
- extraction.rs (50 lines) - Module declarations and re-exports

Benefits:
- No client code changes required (maintains same public API)
- Tests colocated with implementation
- Easier to navigate and maintain
- Follows same pattern as ast/elements/ directory
